PR issues fix
