+++ core-db/src/main/java/jp/co/soramitsu/core_db/AppDatabase.kt
@@ -17,6 +17,7 @@ import jp.co.soramitsu.core_db.dao.NodeDao
 import jp.co.soramitsu.core_db.dao.PhishingAddressDao
 import jp.co.soramitsu.core_db.dao.RuntimeDao
 import jp.co.soramitsu.core_db.dao.StakingRewardDao
+import jp.co.soramitsu.core_db.dao.StakingTotalRewardDao
 import jp.co.soramitsu.core_db.dao.StorageDao
 import jp.co.soramitsu.core_db.dao.TokenDao
 import jp.co.soramitsu.core_db.dao.TransactionDao
@@ -123,4 +124,6 @@ abstract class AppDatabase : RoomDatabase() {
     abstract fun accountStakingDao(): AccountStakingDao
 
     abstract fun stakingRewardsDao(): StakingRewardDao
+
+    abstract fun stakingTotalRewardDao(): StakingTotalRewardDao
 }
