+++ core-db/src/main/java/jp/co/soramitsu/core_db/dao/StakingRewardDao.kt
@@ -17,12 +17,6 @@ abstract class StakingRewardDao {
     @Query("SELECT COUNT(*) FROM staking_rewards WHERE accountAddress = :accountAddress")
     abstract suspend fun rewardCount(accountAddress: String): Int
 
-    @Query("SELECT * FROM total_reward WHERE accountAddress = :accountAddress")
-    abstract fun observeTotalRewards(accountAddress: String): Flow<TotalRewardLocal>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    abstract suspend fun insert(totalRewardLocal: TotalRewardLocal)
-
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     abstract suspend fun insert(rewards: List<StakingRewardLocal>)
 }
