+++ core-db/src/main/java/jp/co/soramitsu/core_db/di/DbApi.kt
@@ -8,6 +8,7 @@ import jp.co.soramitsu.core_db.dao.NodeDao
 import jp.co.soramitsu.core_db.dao.PhishingAddressDao
 import jp.co.soramitsu.core_db.dao.RuntimeDao
 import jp.co.soramitsu.core_db.dao.StakingRewardDao
+import jp.co.soramitsu.core_db.dao.StakingTotalRewardDao
 import jp.co.soramitsu.core_db.dao.StorageDao
 import jp.co.soramitsu.core_db.dao.TokenDao
 import jp.co.soramitsu.core_db.dao.TransactionDao
@@ -35,4 +36,6 @@ interface DbApi {
     fun accountStakingDao(): AccountStakingDao
 
     fun stakingRewardDao(): StakingRewardDao
+
+    fun stakingTotalRewardDao(): StakingTotalRewardDao
 }
