+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/data/network/subscan/request/StakingSumRewardRequest.kt
@@ -2,10 +2,10 @@ package jp.co.soramitsu.feature_staking_impl.data.network.subscan.request
 
 class StakingSumRewardRequest(accountAddress: String) {
     val query = """
-        |{
-        |sumReward
-        |   (id: "$accountAddress")
-        |   {accountTotal}
-        |}
-        |""".trimMargin()
+        {
+        sumReward
+           (id: "$accountAddress")
+           {accountTotal}
+        }
+        """.trimIndent()
 }
