+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/di/StakingFeatureDependencies.kt
@@ -13,6 +13,7 @@ import jp.co.soramitsu.common.validation.ValidationExecutor
 import jp.co.soramitsu.core.storage.StorageCache
 import jp.co.soramitsu.core_db.dao.AccountStakingDao
 import jp.co.soramitsu.core_db.dao.StakingRewardDao
+import jp.co.soramitsu.core_db.dao.StakingTotalRewardDao
 import jp.co.soramitsu.fearless_utils.runtime.RuntimeSnapshot
 import jp.co.soramitsu.fearless_utils.wsrpc.SocketService
 import jp.co.soramitsu.feature_account_api.domain.interfaces.AccountRepository
@@ -66,6 +67,8 @@ interface StakingFeatureDependencies {
     fun accountUpdateScope(): AccountUpdateScope
 
     fun stakingRewardsDao(): StakingRewardDao
+
+    fun stakingTotalRewardsDao(): StakingTotalRewardDao
 
     fun networkApiCreator(): NetworkApiCreator
